# Add locks for de-/compressor + initialize decompressor