@quinnj 2023-05-25
Not sure if @svilupp wasn't able to see this through, but I've cleaned up their code here: https://github.com/apache/arrow-julia/pull/445. I can confirm that I was able to reproduce the original segfault on Apple M1 and that the code in that PR resolves the issue. I believe that PR should also address @baumgold's comments from this PR.